Extracted components: ClientPortalDialog, DateRangeDialog, TrkrefDialog, FilterPreview, SquareRadioGroup
