+++ app/js/components/client_portal_dialog/client_portal_dialog.scss
@@ -1,0 +1,9 @@
+@import '../../../../node_modules/client_portal-assets/dist/sass/colours';
+
+.dialog {
+  &__section-title {
+    color: $light-text-color;
+    font-size: 13px;
+    text-transform: uppercase;
+  }
+}
