+++ app/js/components/dashboard_filters/dashboard_filters.scss
@@ -1,13 +1,5 @@
 @import '../../../../node_modules/client_portal-assets/dist/sass/colours';
 
-.filter{
-  &__title {
-    font-size: 13px;
-    margin: 30px 0;
-    text-transform: uppercase;
-  }
-}
-
 .filter-preview {
   &:hover {
     cursor: pointer;
@@ -27,45 +19,5 @@
     display: flex;
     flex-direction: row;
     padding-left: 20px;
-  }
-}
-
-.flat-radio-group {
-  display: flex;
-  flex-direction: row;
-  list-style: none;
-  padding: 0;
-  position: relative;
-}
-
-.flat-radio-button {
-  &__input {
-    opacity: 0;
-    position: absolute;
-
-    &:checked + .flat-radio-button__label {
-      background-color: $white-two;
-      color: $dark-sky-blue;
-    }
-  }
-
-  &__label {
-    border: 1px solid #ced1d2;
-    font-size: 14px;
-    line-height: 1.3;
-    margin-left: -1px;
-    padding: 11px 20px;
-
-    &:hover {
-      cursor: pointer;
-    }
-  }
-}
-
-.dialog {
-  &__section-title {
-    color: $light-text-color;
-    font-size: 13px;
-    text-transform: uppercase;
   }
 }
