+++ app/js/components/square_radio_group/square_radio_group.scss
@@ -1,0 +1,33 @@
+@import '../../../../node_modules/client_portal-assets/dist/sass/colours';
+
+.square-radio-group {
+  display: flex;
+  flex-direction: row;
+  list-style: none;
+  padding: 0;
+  position: relative;
+}
+
+.square-radio-button {
+  &__input {
+    opacity: 0;
+    position: absolute;
+
+    &:checked + .square-radio-button__label {
+      background-color: $white-two;
+      color: $dark-sky-blue;
+    }
+  }
+
+  &__label {
+    border: 1px solid #ced1d2;
+    font-size: 14px;
+    line-height: 1.3;
+    margin-left: -1px;
+    padding: 11px 20px;
+
+    &:hover {
+      cursor: pointer;
+    }
+  }
+}
